+++ crates/amaru-ledger/src/governance/ratification.rs
@@ -20,7 +20,7 @@ use amaru_kernel::{
     RatificationStatus, StakeCredential, Vote, Voter,
     rational_number::{SafeRatio, into_safe_ratio},
 };
-use amaru_observability::info_span;
+use amaru_observability::{debug_span, info_span};
 use num::Zero;
 use tracing::{Span, field};
 
@@ -282,7 +282,11 @@ impl<'distr> RatificationContext<'distr> {
     }
 
     fn new_ratify_span(id: &ProposalId, proposal: &ProposalEnum) -> Span {
-        info_span!(ledger::governance::RATIFYING, proposal_id = id.to_string(), proposal_kind = proposal.display_kind())
+        debug_span!(
+            ledger::governance::RATIFYING,
+            proposal_id = id.to_string(),
+            proposal_kind = proposal.display_kind()
+        )
     }
 
     fn is_accepted_by_everyone(